Committed a simplified version of flattening in Agda.
I've tried to construct a simpler version where I can clearly see why we
need to deal with environment properly.

The main difference is that flattening directly interprets source
expressions, whereas in the original file we had tried to construct
an abstract syntax of flattened expressions as an intermediate step.

There are several examples in the end, which we run using the original
and the flattened interpretation. The last example exposes a situation
where we have to deal with the captured environment of a lifted
function.