Nearing completion. Only one major hole left.
Ok, I think I understand most of the issues involved (finally).

Hole number 3 is the most critical. It corresponds to expandPA in
the HtM paper. I could use an example that actually needs it.

The current version is deliberately close to the HtM paper. I've
replaced all "shapes" by lengths. The lengths are the other critical
point though. There's a lot of type unsafety arising from the fact
that we can incorrectly pass lengths.

I think a next step would be to keep track of the lengths of all
the lists involved at compile time. This is currently simple for
our source language, as the only list operations are constants
(for which we trivially know the length) and map (which preserves
length).