Bugfix in the Reuse introduction in the X100 code generator: Only replace a child by a label reference if the parent is not already a Reuse operator.
Only insert a label reference if the parent is not a Reuse operator.
Otherwise, a shared node would be visited from the same parent again
after being equipped with a Reuse operator at the first visit. During
the second visit, the node would be replaced by a label reference in
the reuse operator, leading to a self-reference.